Refactor DiagnosticAnalyzer runner to a helper class
